Rename Types -> Data
