+++ src/ShellRun/Data/Command.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE DerivingVia #-}
 
 -- | Provides the 'Command' wrapper for commands.
-module ShellRun.Types.Command (Command (..)) where
+module ShellRun.Data.Command (Command (..)) where
 
 import Data.Text (Text)
