+++ src/ShellRun/Data/Env.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE MultiParamTypeClasses #-}
 
 -- | Provides core 'Env' types.
-module ShellRun.Types.Env
+module ShellRun.Data.Env
   ( -- * \"HasX\" style typeclasses required for our concrete Env type.
     HasCommandDisplay (..),
     HasLogQueue (..),
