+++ src/ShellRun/Parsing/Legend.hs
@@ -11,7 +11,7 @@ import Control.Exception qualified as Ex
 import Data.Text (Text)
 import Data.Text qualified as T
 import ShellRun.Parsing.Legend.Internal qualified as Internal
-import ShellRun.Types.Legend (LegendErr (..), LegendMap)
+import ShellRun.Data.Legend (LegendErr (..), LegendMap)
 
 -- | Given a filepath, attempts to parse the file's contents into
 -- a 'LegendMap'. If the file does not exist or the parsing fails
