+++ src/ShellRun/Utils.hs
@@ -24,8 +24,8 @@ where
 import Data.Text (Text)
 import ShellRun.Math (NonNegative)
 import ShellRun.Math qualified as Math
-import ShellRun.Types.Command (Command (..))
-import ShellRun.Types.Env (CommandDisplay (..))
+import ShellRun.Data.Command (Command (..))
+import ShellRun.Data.Env (CommandDisplay (..))
 import ShellRun.Utils.Internal qualified as UtilsI
 import ShellRun.Utils.Text as TextUtils
 import System.Clock (TimeSpec (..))
