+++ test/integration/MockShell/BadLegendMockShell.hs
@@ -11,7 +11,7 @@ import MockEnv (MockEnv)
 import MockShell.MockShellBase (MockShellBase (..))
 import ShellRun.Class.MonadShell (MonadShell (..))
 import ShellRun.Logging (MonadLogger (..))
-import ShellRun.Types.Legend (LegendErr (..))
+import ShellRun.Data.Legend (LegendErr (..))
 
 -- | 'BadLegendMockShell' is intended to test a run of
 -- 'ShellRun.runShell' when the path to the legend file is bad.
