+++ test/unit/Props/ShellRun/Parsing/Commands.hs
@@ -16,8 +16,8 @@ import Hedgehog qualified as H
 import Hedgehog.Gen qualified as Gen
 import Hedgehog.Range qualified as Range
 import ShellRun.Parsing.Commands qualified as ParseCommands
-import ShellRun.Types.Command qualified as Command
-import ShellRun.Types.Legend (LegendMap)
+import ShellRun.Data.Command qualified as Command
+import ShellRun.Data.Legend (LegendMap)
 import Test.Tasty (TestTree)
 import Test.Tasty qualified as T
 import Test.Tasty.Hedgehog qualified as TH
