+++ test/unit/Props/ShellRun/Parsing/Legend/Internal.hs
@@ -14,7 +14,7 @@ import Hedgehog qualified as H
 import Hedgehog.Gen qualified as Gen
 import Hedgehog.Range qualified as Range
 import ShellRun.Parsing.Legend.Internal qualified as Internal
-import ShellRun.Types.Legend (LegendMap)
+import ShellRun.Data.Legend (LegendMap)
 import Test.Tasty (TestTree)
 import Test.Tasty qualified as T
 import Test.Tasty.Hedgehog qualified as TH
