+++ test/unit/Specs/ShellRun/Parsing/Commands.hs
@@ -6,8 +6,8 @@ module Specs.ShellRun.Parsing.Commands (specs) where
 
 import Data.Map.Strict qualified as Map
 import ShellRun.Parsing.Commands qualified as ParseCommands
-import ShellRun.Types.Command (Command (..))
-import ShellRun.Types.Legend (LegendErr (..), LegendMap)
+import ShellRun.Data.Command (Command (..))
+import ShellRun.Data.Legend (LegendErr (..), LegendMap)
 import Test.Hspec (shouldBe)
 import Test.Hspec qualified as Hspec
 import Test.Tasty (TestTree)
