+++ test/unit/Specs/ShellRun/Parsing/Legend/Internal.hs
@@ -6,7 +6,7 @@ module Specs.ShellRun.Parsing.Legend.Internal (specs) where
 
 import Data.Map.Strict qualified as Map
 import ShellRun.Parsing.Legend.Internal qualified as Internal
-import ShellRun.Types.Legend (LegendErr (..))
+import ShellRun.Data.Legend (LegendErr (..))
 import Test.Hspec (shouldBe)
 import Test.Hspec qualified as Hspec
 import Test.Tasty (TestTree)
